requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
pyTelegramBotAPI==4.14.0
asyncpg==0.29.0
aiohttp==3.9.1
python-dotenv==1.0.0
websockets==12.0